Revert "Moved ContentPreparer to posts directory"
This reverts commit d4afffd12706e6b591ab9cb267e16119845f39d9.
